feat: Create cache for inventory structure query
